+++ app/src/main/java/com/example/chapmac/rakkan/mqtt_app_test/TabActivity.java
@@ -47,8 +47,8 @@ public class TabActivity extends AppCompatActivity {
         fab.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
+                SubscribeDialog subscribeDialog = new SubscribeDialog();
+                subscribeDialog.show(getSupportFragmentManager(),"Subscribe Dialog");
             }
         });
