+++ app/src/main/res/layout/fragment_subscribe.xml
@@ -36,7 +36,6 @@
             android:id="@+id/recyclerView"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:padding="4dp"
             android:scrollbars="vertical"
 
             />
