+++ app/src/main/res/layout/subscribe_item.xml
@@ -6,8 +6,7 @@
 
     <RelativeLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_margin="4dp">
+        android:layout_height="match_parent">
 
         <ImageView
             android:id="@+id/imageView"
@@ -20,12 +19,11 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_alignParentTop="true"
-            android:layout_toEndOf="@+id/imageView"
-            android:layout_marginTop="6dp"
             android:layout_marginStart="8dp"
+            android:layout_marginTop="6dp"
+            android:layout_toEndOf="@+id/imageView"
             android:text="Line1"
-
-             />
+            />
 
         <TextView
             android:id="@+id/textView1"
@@ -34,8 +32,7 @@
             android:layout_below="@+id/textView"
             android:layout_marginStart="8dp"
             android:layout_toEndOf="@+id/imageView"
-            android:text="Line2"
-             />
+            android:text="Line2" />
 
     </RelativeLayout>
